+++ pages/manageFormula.php
@@ -5,6 +5,7 @@ require_once(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/config.php');
 require_once(__ROOT__.'/inc/opendb.php');
 require_once(__ROOT__.'/inc/settings.php');
+require_once(__ROOT__.'/inc/product.php');
 require_once(__ROOT__.'/func/labelMap.php');
 require_once(__ROOT__.'/func/get_formula_notes.php');
 
@@ -362,6 +363,8 @@ if($_POST['action'] == 'addFormula'){
 	}else{
 		if(mysqli_query($conn, "INSERT INTO formulasMetaData (fid, name, notes, profile, catClass, finalType, customer_id) VALUES ('$fid', '$name', '$notes', '$profile', '$catClass', '$finalType', '$customer_id')")){
 			$last_id = mysqli_insert_id($conn);
+			$fullver = $product.' '.$ver;
+			mysqli_query($conn, "INSERT INTO formulasTags (formula_id, tag_name) VALUES ('$last_id','$fullver')");
 			$response = array(
 				"success" => array(
 				"id" => (int)$last_id,
@@ -394,6 +397,7 @@ if($_POST['action'] == 'delete' && $_POST['fid']){
 		mysqli_query($conn, "DELETE FROM formulasMetaData WHERE fid = '$fid'");
 		mysqli_query($conn, "DELETE FROM formulasRevisions WHERE fid = '$fid'");
 		mysqli_query($conn, "DELETE FROM formula_history WHERE fid = '".$meta['id']."'");
+		mysqli_query($conn, "DELETE FROM formulasTags WHERE formula_id = '".$meta['id']."'");
 		$response['success'] = 'Formula '.$fname.' deleted!';
 	}else{
 		$response['error'] = 'Error deleting '.$fname.' formula!';