+++ releasenotes.md
@@ -2,3 +2,4 @@ Whats New in v6.5
 ----------------------
 - You can now merge similar ingredients in formula view to one and sum up their quantities
 - Rewrite ingredient replacement dialog
+- We auto add a tag with PV Version when you create a new formula